#1036: pare down engine entrypoint even more
